Fix code smells (#2)
* Fix code smells

* Update submodules

* Configure sonar-scan action for submodules

* Configure sonar-scan action for submodules

* Configure sonar-scan action for submodules

* Configure sonar-scan action for submodules

* Configure sonar-scan action for submodules

* Configure sonar-scan action for submodules

* Configure sonar-scan action for submodules

* Update scripts/sonar_scan.sh and run_docker_sonar_scan.sh

* Update run_docker_fw_build.sh

* Run sonar-scan tasks in parallel

* Run sonar-scan tasks in parallel

+++ components/embedded-i2c-sen66-master/sensirion_i2c_hal.c
@@ -83,7 +83,7 @@ void sensirion_i2c_hal_free(void) {
  * @param count   number of bytes to read from I2C and store in the buffer
  * @returns 0 on success, error code otherwise
  */
-int8_t sensirion_i2c_hal_read(uint8_t address, uint8_t* data, uint8_t count) {
+int16_t sensirion_i2c_hal_read(uint8_t address, uint8_t* data, uint8_t count) {
     /* TODO:IMPLEMENT */
     return NOT_IMPLEMENTED_ERROR;
 }
@@ -99,8 +99,8 @@ int8_t sensirion_i2c_hal_read(uint8_t address, uint8_t* data, uint8_t count) {
  * @param count   number of bytes to read from the buffer and send over I2C
  * @returns 0 on success, error code otherwise
  */
-int8_t sensirion_i2c_hal_write(uint8_t address, const uint8_t* data,
-                               uint8_t count) {
+int16_t sensirion_i2c_hal_write(uint8_t address, const uint8_t* data,
+                                uint8_t count) {
     /* TODO:IMPLEMENT */
     return NOT_IMPLEMENTED_ERROR;
 }
